Script to categorize multiple docs. Move AI conversation into external function.
