+++ functions-v2/lib/src/ai-categorize-document.ts
@@ -1,0 +1,66 @@
+import OpenAI from "openai";
+import {zodResponseFormat} from "openai/helpers/zod";
+import {z} from "zod";
+import fs from "node:fs/promises";
+
+const prompt = `This is a picture of a student document.
+They are working on engineering task. Please tell me which of the following areas of their design they are focusing on:
+- user: who's it for?
+- environment: where's it used?
+- form: what's it look like?
+- function: what does it do?
+and why you chose that area.
+Or if the document doesn't include enough content to clearly identify a focus area let me know by setting "success" to
+"false" and "category" to "unknown".
+Your answer should be a JSON document in the given format.`;
+
+// Require a specific JSON schema for the model output
+const CategorizationResponse = z.object({
+  success: z.boolean(
+    {description: "Whether a category could be determined from the input"}),
+  category: z.enum(["user", "environment", "form", "function", "unknown"],
+    {description: "The focus area of the document"}),
+  keyIndicators: z.array(z.string(),
+    {description: "List of main features or elements of the document that support this categorization"}),
+  discussion: z.string(
+    {description: "Any other relevant information."}),
+});
+
+export default async function categorizeDocument(file: string) {
+  const imageLoading = fs.readFile(file).then((data) => data.toString("base64"));
+  const image = await imageLoading;
+
+  const openai = new OpenAI({apiKey: process.env.OPENAI_API_KEY});
+  try {
+    return openai.beta.chat.completions.parse({
+      // model: "gpt-4o-mini",
+      model: "gpt-4o-2024-08-06",
+      messages: [
+        {
+          role: "system",
+          content: "You are a teaching assistant in an engineering design course.",
+        },
+        {
+          role: "user",
+          content: [
+            {
+              type: "text",
+              text: prompt,
+            },
+            {
+              type: "image_url",
+              image_url: {
+                url: `data:image/png;base64,${image}`,
+                detail: "auto", // auto, low, high
+              },
+            },
+          ],
+        },
+      ],
+      response_format: zodResponseFormat(CategorizationResponse, "categorization-response"),
+    });
+  } catch (error) {
+    console.log("OpenAI error", error);
+    return undefined;
+  }
+}
